insomnia/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
langgraph>=0.2.0
mcp>=1.0.0
openai>=1.40.0
python-dotenv>=1.0.1
rich>=13.7.0
requests